Split defer-bundle publish check tests under the file-size budget.
Move the new runRepoChecks defer-bundle cases into a sibling node test
file so checks.node.test.ts stays at or below 2000 lines.
